Move web service and persistence to other files.
