+++ CodingPractice/ComposableArchitecture/WebServiceEnvironment.swift
@@ -1,0 +1,14 @@
+import ComposableArchitecture
+import Foundation
+
+struct WebServiceEnvironment {
+    var getProducts: () -> Effect<[Product], WebServiceError>
+}
+
+extension WebServiceEnvironment {
+    static func makeLive(webServiceClient: WebServiceClient) -> Self {
+        return .init {
+            webServiceClient.getProducts()
+        }
+    }
+}
